add and mod funs

+++ f.download.R
@@ -3,7 +3,7 @@
 #' Downloads a large number of files with a wide range of options, including custom sleep times, custom retries, custom sleep during retries, timeout and a debug toggle for testing a reduced number of files. Will check the filenames in the destination folder to ensure that only files scheduled to be downloaded are present. Will skip already downloaded files.
 
 
-#' @param url Character. A vector of URLs. Will skip "NA" values.
+#' @param url Character. A vector of URLS.
 #' @param filename Character. A vector of filenames.
 #' @param dir Character. The destination directory. Will be created if not already present.
 #' @param sleep.min Positive Integer. Minimum number of seconds to randomly sleep between requests.
@@ -11,7 +11,6 @@
 #' @param retries Positive Integer. Number of retries for entire download set.
 #' @param retry.sleep.min Positive Integer. Minimum number of seconds to randomly sleep between requests in retry mode.
 #' @param retry.sleep.max Positive Integer. Maximum number of seconds to randomly sleep between requests in retry mode.
-#' @param random.order Logical. Whether to download files in random order. Defaults to TRUE.
 #' @param timeout Positive Integer. Number of seconds for a download to timeout, even active ongoing ones.
 #' @param debug.toggle Logical. Whether to activate debugging mode.
 #' @param debug.files Positive Integer. The number of files to download during debugging mode. Default is 500.
@@ -19,21 +18,26 @@
 #' @return Character. A vector of the downloaded file names.
 
 
+
 f.download <- function(url,
                        filename,
                        dir,
+                       clean = TRUE,
                        sleep.min = 0,
                        sleep.max = 0.1,
                        retries = 3,
                        retry.sleep.min = 2,
                        retry.sleep.max = 5,
-                       random.order = TRUE,
                        timeout = 300,
                        debug.toggle = FALSE,
                        debug.files = 500){
 
+
+    
     ## Set timeout for downloads
     options(timeout = timeout)
+
+    
     
     ## Test for equality of length
     if(length(url) != length(filename)){
@@ -42,10 +46,13 @@ f.download <- function(url,
         
     }
 
-    ## Create folder
-    dir.create(dir, showWarnings = FALSE)
     
-    ## Create Data Frame
+    ## Create download folder
+    dir.create(dir, showWarnings = FALSE, recursive = TRUE)
+
+
+    
+    ## Create data frame
     df <- data.frame(url, filename)
 
 
@@ -59,10 +66,14 @@ f.download <- function(url,
 
     
     ## Clean folder: Only files included in 'filename' may be present
-    
-    files.all <- list.files(dir, full.names = TRUE)
-    delete <- setdiff(files.all, file.path(dir, df$filename))
-    unlink(delete)
+
+    if (clean == TRUE){
+        
+        files.all <- list.files(dir, full.names = TRUE)
+        delete <- setdiff(files.all, file.path(dir, df$filename))
+        unlink(delete)
+
+    }
 
 
     
@@ -77,32 +88,14 @@ f.download <- function(url,
     
     ## Download: First Try
 
-    if(random.order == TRUE){
-        
-        download.order <- sample(nrow(df.todo))
-
-    }else{
-        
-        download.order <- 1:nrow(df.todo)
-        
-    }
-
 
     if(nrow(df.todo) > 0){
         
-        for (i in download.order){
-            
-            tryCatch({download.file(url = df.todo$url[i],
-                                    destfile = file.path(dir,
-                                                         df.todo$filename[i]))
-            },
-            
-            error = function(cond) {
-                return(NA)}
-            )
-            
-            Sys.sleep(runif(1, sleep.min, sleep.max))
-        }
+        result.todo <- mapply(f.download_robust,
+                              url = df.todo$url,
+                              destfile = file.path(dir, df.todo$filename),
+                              sleep.min = sleep.min,
+                              sleep.max = sleep.max)
 
     }
     
@@ -112,42 +105,24 @@ f.download <- function(url,
 
     for(i in 1:retries){
         
-        ## Missing files
-        
         files.present <- list.files(dir)
         filename.missing <- setdiff(df$filename, files.present)
         df.missing  <- df[df$filename %in% filename.missing,]
         df.missing <- df.missing[!is.na(df.missing$url),] # skip NA urls
 
+
         if(nrow(df.missing) > 0){
-
-
-            if(random.order == TRUE){
-                
-                download.order <- sample(nrow(df.missing))
-
-            }else{
-                
-                download.order <- 1:nrow(df.missing)
-                
-            }
             
-            
-            for (i in download.order){
-                
-                tryCatch({download.file(url = df.missing$url[i],
-                                        destfile = file.path(dir,
-                                                             df.missing$filename[i]))
-                },
-                error = function(cond) {
-                    return(NA)}
-                )     
-                
-                Sys.sleep(runif(1, retry.sleep.min, retry.sleep.max))
-            } 
+            result.retry <- mapply(f.download_robust,
+                                   url = df.missing$url,
+                                   destfile = file.path(dir, df.missing$filename),
+                                   sleep.min = retry.sleep.min,
+                                   sleep.max = retry.sleep.max)
+
         }
-        
+
     }
+    
 
 
 
@@ -169,4 +144,41 @@ f.download <- function(url,
 
     return(files.all)
     
+}
+
+
+
+
+
+
+
+#' Wrapper function for download.file that returns 'NA' instead of an error when a file is not available.
+#'
+#' @param url A valid URL.
+#' @param destfile The destination file.
+#' @param sleep.min Positive Integer. Minimum number of seconds to randomly sleep between requests.
+#' @param sleep.max Positive Integer. Maximum number of seconds to randomly sleep between requests.
+#'
+#' @return Character. Returns destination file invisibly.
+
+
+
+
+f.download_robust <- function(url,
+                              destfile,
+                              sleep.min = 0,
+                              sleep.max = 0){
+
+    tryCatch({download.file(url = url,
+                            destfile = destfile)
+    },
+    error = function(cond) {
+        return(NA)}
+
+    )
+
+    Sys.sleep(runif(1, sleep.min, sleep.max))
+
+    invisible(destfile)
+
 }
